feat(tests): add 'test_containers' and move some memdata tests
